modified : restore lazy load setting
Because lazy.nvim cache(?) is so fast,
there does not exist big difference in load time between with and without lazy
loading unless you need to it only for specific file types.

It will be no different in spite of first time to load according to
computer environment

I usually use Oil browser at startup neovim. and it invokes event
'BufReadPost'. if we use lazy loading at BufReadPost, the first time
to load oil browser has delay to open. it is very inconvenient.
So I set 'lazy' to false in other plugins which must be not needed to
lazy loading.