fix(markdown) : adjust `left_pad` to system default value
I set tabstop to 4 as default and left_pad of markdown is 2.
It made that first and second indentaiton is differnet value.
first indentation is 2 because tabstop 4 - padding 2
second indenation is 8 because tabstop 4*2 and padding doesn't works.

So I recognize it is better to set left_pad to default tapstop value.