feat(options) : remove shada.tmp file after VimLeave
- autocmd for removing shada tmp file
When VimLeave, it must be remove shada.tmp file but it doesn't.
It is very chronic issue of neovim. so remove the shada file after
`main.shada` is saved.

- autocmd to taskkill harper-ls
sometimes harper-ls cannot be terminated properly after VimLeave.
So I want to terminate the process after VimLeave, but the way with name
has some problem that it will terminate other normal process of
harper-ls.

this solution is on hold