fix(markdown) : rollback autolistNewBullet shortcut
I added <Esc> and <Right> after AutolistNewBullet() to change cursor
location to beginning of the line when I cut a line in the middle of
them and make new bullet.

But this makes some bug when I enter <CR> at the end of bullet line. it
cannot maintain the cursor location