fix(autocmds) : comment autocmd for erase harper-ls.exe at BufreadPre
- Problem
The system call using lua api in Windows is very slow again.
Whenever `BufReadPre`, check the process id of nvim and harper,
and taskkill them for memory management and removing unused harper-ls

But it makes Buffer loading time slower, and it is the main reason for
longer UIEnter time.

- Solution
This chunk is commented. I'll find some other solution in the furture.