fix(options) : remove set clipboard provider
At #fbc7e1d, I forced set clipboard provider because It is slow to copy
and paste from system clipboard at first time. I think it is the time to
search where the clipboard provider is, but It seems not.

`:checkhealth` said win32yank is detected always and this problem
invokes continuously
So this chunk is removed