fix(mason) : simplify about mason tools
Using `vim.lsp,config()` in neovim v0.11, `nvim-lspconfig` doesn't be
needed. so the `mason-tool-installer` and `mason-lspconfig` don't be
needed also. so this plugins are removed.

only `mason.nvim` is alive as lsp and formatter installer.
it is lazy loaded by cmd `:Mason` when I want to install the tools.
the lsp server's paths is registered in core/lsp.lua