fix(snacks) : include all recent files without filters
1. Problem

snacks.picker.recent() shows only files in out of `nvim-data/`.
Files in `nvim-data/` are hidden automatically.

2. Reason
I don't know what is the exact reason, but I guess it because of default
filter setting of recent() function.
```lua
    filter = {
	paths = {
	  [vim.fn.stdpath("data")] = false,
	  [vim.fn.stdpath("cache")] = false,
	  [vim.fn.stdpath("state")] = false,
	},
      },
```

It makes files in stdpath('data') are filtered automatically.

3. Solution

After makes the value of `paths` to false, it works as what I expect.