fix(yazi) : set cwd() for <leader>ee
1. Problem

if yazi api has no path to parameter, it recognize the path is cwd to
open yazi.
But somedays, it invoke some bugs and cannot open any window.

2. solution

It has a bug yet, getcwd() is transferred to the parameter of yazi api
to open current directory.
But there are some display bugs in specific directories yet.