feat(utils) : add smart_fold() function and key mapping
default folding key mapping is separated with function.
Separating to toggle(za) / open(zo) / close(zc) / create(zf) / open nested fold(zv) is annoying.
and It cannot recognize treesitter. I must set visual region to create it.

If fold is created already, it just toggle the fold.
If fold is not created, make it.
To make smarter behavior, it detects 'comment', 'string(docstring in
python)' and create fold in normal mode. Detection region is continuesly adjacent
treesitter node. It it is in visual mode, it follows 'zf' behavior