feat(smart_cr) : add `smart_enter()` keymap for parentheses <CR>
1. Problem

Some filetypes are has their `indenexpr` and it could do proper behavior
for someone but not me. I always use <CR> in parentheses to make
indented cursor between parentheses like this.
a = {
|
}

Especially, python has additional wide indent, and nvim-autopairs has
its configuration using keycodes in map_cr(), it would work in lua and matlab, but
it doesn't work in python. that's what I concerned.

2. Solution

Remove All other map_cr() of filetype in nvim-autopairs configuration.
and Create my own function to check cursor is inside of parentheses, If
then, insert indent when <CR> is input.

It works for python and nvim-autopairs, but it has some problem yet.
1) inserting indent and new line is must does when there are no word in
parentheses.
2) some filetypes has dedicated <CR> keymap from plugins. it needs to
consider fusion with them