fix(snacks) : use myown oldfile picker
Default recent() of snacks picker is weird.
It omits in nvim-data/ file and if there are duplicated file name, only
show one file although they are two file which is in different path.
And It doesn't follow order of vim.v.oldfiles which means the order is
by how recently the file was opened.

My purpose is
1) show file list by order of vim.v.oldfiles even in a search situation.
2) show files in any directory includes nvim-data/.
3) omit ignored patterns what I want