fix(utils) : truncate directory of oldfiles_picker() if it's too long
Length of some file path are too long to be displayed in
oldfiles_picker(). It is annoying.

While item.text remains to search, we made one additional variable that named
`diraliase` which means that dirname displayed in picker().

If text is too long, directories are truncated from start of dirname.
but it can be searchable with removed directory because text is remained