fix : resolve hidden fenced_code_block marker in markdown issue.
It is the final solution to show code_block marker without modifying
query of nvim-treesitter.

If markview.nvim path is prepended to rtp in configuration block of
markview, It cannot render properly when I open some file using picker.
It is ok when I open .md in oil.nvim but snacks picker doesn't.

To solve this, we need to set `lazy=false` to markview.nvim.
but loading time is 24ms in Windows... very slow.
so We moved the rtp prepending section to nvim-treesitter's one to
recognize the markdown query before markview.nvim load.

I think If I set lazy load using `ft = 'markdown'`, the query executes
first before plugin is loaded. It seems the reason as I think.

Now All done.