fix : bug fix 6772cc1,
From 6772cc1, I thought it is the best config to prevent hiding
fenced_code_block mark.
But it isn't. If the .md is removed but remain in buffer list with
unload state, It will show code block mark.
But If the .md is wiped and don't be remained in buffer list, marker is
hidden. this solution must need `lazy = true` to `markview.nvim`
configuration.

But I found the real solution.
After all lazy.nvim plugins are loaded, then add `markview.nvim` path to
rtp. I implemented this in lsp.lua which is executed after lazy.nvim
finished.

It works fine and It is same with modifying query of nvim-treesitter
directly.